+++ EliteAPI.Events/Station/EngineerCraftEvent.cs
@@ -33,7 +33,7 @@ public readonly struct EngineerCraftEvent : IEvent
     public string BlueprintName { get; init; }
 
     [JsonProperty("Level")]
-    public int Level { get; init; }
+    public long Level { get; init; }
 
     [JsonProperty("Quality")]
     public double Quality { get; init; }
