+++ EliteAPI.Events/Travel/DockedEvent.cs
@@ -63,13 +63,13 @@ public readonly struct DockedEvent : IEvent
     public readonly struct LandingPadsInfo
     {
         [JsonProperty("Small")]
-        public int Small { get; init; }
+        public long Small { get; init; }
 
         [JsonProperty("Medium")]
-        public int Medium { get; init; }
+        public long Medium { get; init; }
 
         [JsonProperty("Large")]
-        public int Large { get; init; }
+        public long Large { get; init; }
     }
